fix(cc1101): drain RX FIFO tail on EOP and clarify GDO2 docs
Address PR review feedback on the GDO2 FIFO threshold feature:

- cc1101.cpp: with GDO2 wired, the RX drain loop could exit on GDO0
  end-of-packet while GDO2 was LOW, leaving a sub-threshold tail unread
  in the FIFO. Add a final drain after the loop (with overflow and
  buffer-bounds checks) so packets are fully drained before parsing.
- cc1101.h: update cc1101_set_gdo2_pin() doc comment to describe the
  dynamic per-phase IOCFG2 behaviour (TX 0x02 threshold, RX 0x01
  threshold/EOP), the SPI-polling fallback, and wiring guidance.
- wifi_serial.h/.cpp: cast the head/tail ring-buffer delta to uint16_t
  before masking so wraparound is well-defined instead of relying on
  signed int promotion.

+++ ESPHOME-release/everblu_meter/wifi_serial.h
@@ -74,8 +74,11 @@ private:
     // Returns number of free bytes in the ring buffer
     inline uint16_t _free() const
     {
+        // Cast the head/tail delta to uint16_t before masking: the subtraction is
+        // otherwise promoted to signed int and can be negative after wraparound,
+        // which would make the bitwise AND rely on implementation-defined behaviour.
         return (WIFI_SERIAL_TX_BUF_SIZE - 1) -
-               ((_head - _tail) & (WIFI_SERIAL_TX_BUF_SIZE - 1));
+               ((uint16_t)(_head - _tail) & (WIFI_SERIAL_TX_BUF_SIZE - 1));
     }
 
     // Enqueue one byte; returns false and increments _dropped if full
